got basic things working, moving onto style
