+++ typescript/packages/spellbookjr/app/recipes/[hash]/page.tsx
@@ -1,13 +1,17 @@
 import { getBlobByHash, getBlobScreenshotUrl } from "@/app/lib/blobby";
 import { notFound } from "next/navigation";
 import RecipeDetail from "@/app/components/recipe-detail";
-import JsonView from "@uiw/react-json-view";
+
 interface RecipeDetailProps {
   params: { hash: string };
 }
 
 export default async function RecipeDetailPage({ params }: RecipeDetailProps) {
-  const hash = params.hash as string;
+  const { hash } = await params;
+
+  if (!hash) {
+    notFound();
+  }
 
   let blob;
   try {
